requirements.txt
alpaca-trade-api>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
textblob>=0.15.3
yfinance>=0.2.3
schedule>=1.1.0
ta>=0.10.2
alpha_vantage>=2.3.1
websocket-client>=1.6.1
plotly>=5.13.0
scipy>=1.10.0
scikit-learn>=1.2.2
python-telegram-bot>=20.3
snscrape==0.7.1
textblob==0.17.1
